+++ test/service.test.js
@@ -47,6 +47,10 @@ class FakeStore {
     }];
   }
 
+  searchText() {
+    return [];
+  }
+
   close() {}
 }